[CPDLP-1854] Remove ProfileAttributes validations dependency
