+++ app/controllers/api/v1/npq_participants_controller.rb
@@ -42,6 +42,12 @@ module Api
         render_from_service(service, serializer_class, params: { cpd_lead_provider: current_user })
       end
 
+      def change_schedule
+        service = ChangeSchedule.new(params_for_recorder)
+
+        render_from_service(service, serializer_class, params: { cpd_lead_provider: current_user })
+      end
+
     private
 
       def serializer_class
