+++ app/services/participants/profile_attributes.rb
@@ -6,12 +6,13 @@ module Participants
     include ActiveModel::Validations
 
     included do
-      attr_accessor :course_identifier, :participant_id, :cpd_lead_provider, :force_training_status_change
+      attr_accessor :course_identifier, :participant_id, :cpd_lead_provider
+      attr_accessor :force_training_status_change
 
       validates :course_identifier, course: true, presence: { message: I18n.t(:missing_course_identifier) }
       validates :participant_id, presence: { message: I18n.t(:missing_participant_id) }
-      validates :participant_id, format: { with: /\A[0-9a-f]{8}-[0-9a-f]{4}-[1-5][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}\Z/, message: I18n.t("errors.participant_id.invalid") }, allow_blank: true
       validates :cpd_lead_provider, presence: { message: I18n.t(:missing_cpd_lead_provider) }
+      validates :participant_id, format: /\A[0-9a-f]{8}-[0-9a-f]{4}-[1-5][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}\Z/, allow_blank: true
       validate :course_valid_for_participant
       validate :participant_has_user_profile
     end
@@ -30,10 +31,6 @@ module Participants
 
     def participant_identity
       @participant_identity ||= ParticipantIdentity.find_by(external_identifier: participant_id)
-    end
-
-    def relevant_induction_record_for_profile(participant_profile)
-      participant_profile.relevant_induction_record(lead_provider:)
     end
 
   private
