+++ app/services/valid_courses_resolver.rb
@@ -16,14 +16,20 @@ class ValidCoursesResolver
   end
 end
 
+class CourseValidator < ActiveModel::Validator
+  def validate(record)
+    return if has_profile_for_course_given_course_identitfier?(record)
 
-def valid_courses
-  case course_identifier
-  when "ecf-induction"
-    %w[ecf-induction]
-  when "ecf-mentor"
-    %w[ecf-mentor]
-  else
-    errors.add(:course_identifier, I18n.t(:invalid_identifier))
+    record.errors.add(:course_identifier, I18n.t(:invalid_course))
+  end
+
+private
+
+  def has_profile_for_course_given_course_identitfier?(record)
+    return unless record.participant_identity&.user
+
+    record.participant_identity.user.participant_profiles.active_record.any? do |participant_profile|
+      participant_profile.class::COURSE_IDENTIFIERS.include?(record.course_identifier)
+    end
   end
 end
